refactor: keep a single common MembershipService interface
The previous commit split MembershipService into Oidc/Basic-specific
interfaces. Reconsidered: callers (converters, bean configs, tests) want
to inject a single common type, and Spring DI already resolves to
exactly one impl per Spring context via the auth-method conditional.

Collapse back to one MembershipService interface with a single abstract
factory method createProvider(principalId, type, claims) — every
principal has an ID, a type (USER/CLIENT), and a (possibly empty) claims
map. A named convenience createProviderForUser(principalId) delegates
with PrincipalType.USER and an empty claims map; BASIC callers use it
directly without spelling out PrincipalType.USER or Map.of() at the
call site. PrincipalType moves back to MembershipService.

The returned MembershipProvider (renamed from MembershipResolver) is a
lazy, memoised view of a principal's memberships — `provider.groups()`
reads naturally and matches how the values are wired as suppliers on
the CamundaAuthentication builder.

Impls drop the redundant Default-prefix (there is only one impl per
auth method per secondary-storage state): DefaultOidcMembershipService
-> OidcMembershipService, DefaultBasicMembershipService ->
BasicMembershipService, NoDbOidcMembershipService unchanged.

WebSecurityOidcTestContext now carries @ConditionalOnAuthenticationMethod
(OIDC) so the broad component-scan in BasicAuthWebSecurityConfig
ParameterizedTest no longer registers an OIDC-flavoured MembershipService
bean in BASIC mode.
