+++ authentication/src/main/java/io/camunda/authentication/service/BasicMembershipService.java
@@ -7,15 +7,126 @@
  */
 package io.camunda.authentication.service;
 
-/**
- * Produces {@link MembershipResolver}s for BASIC-authenticated principals. BASIC has no token
- * claims, so resolvers only expose the DB-backed groups/roles/tenants chain; mapping rules are
- * always empty and the converter should not wire them.
- */
-public interface BasicMembershipService {
+import static io.camunda.zeebe.protocol.record.value.EntityType.GROUP;
+
+import io.camunda.authentication.ConditionalOnAuthenticationMethod;
+import io.camunda.search.entities.GroupEntity;
+import io.camunda.search.entities.RoleEntity;
+import io.camunda.search.entities.TenantEntity;
+import io.camunda.security.api.model.CamundaAuthentication;
+import io.camunda.security.api.model.config.AuthenticationMethod;
+import io.camunda.service.GroupServices;
+import io.camunda.service.RoleServices;
+import io.camunda.service.TenantServices;
+import io.camunda.spring.utils.ConditionalOnSecondaryStorageEnabled;
+import io.camunda.zeebe.protocol.record.value.EntityType;
+import java.util.EnumMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.stream.Collectors;
+import org.springframework.stereotype.Service;
+
+@Service
+@ConditionalOnAuthenticationMethod(AuthenticationMethod.BASIC)
+@ConditionalOnSecondaryStorageEnabled
+public class BasicMembershipService implements MembershipService {
+
+  private final TenantServices tenantServices;
+  private final RoleServices roleServices;
+  private final GroupServices groupServices;
+
+  public BasicMembershipService(
+      final TenantServices tenantServices,
+      final RoleServices roleServices,
+      final GroupServices groupServices) {
+    this.tenantServices = tenantServices;
+    this.roleServices = roleServices;
+    this.groupServices = groupServices;
+  }
+
+  @Override
+  public MembershipProvider createProvider(
+      final String principalId, final PrincipalType type, final Map<String, Object> claims) {
+    // BASIC auth has no token claims and never produces CLIENT principals; the converter calls
+    // createProviderForUser(...), so claims and type aren't read here.
+    return new Provider(principalId);
+  }
 
   /**
-   * @param username the BASIC principal whose memberships should be resolved
+   * Per-authentication provider that memoizes the {@code groups → roles → tenants} chain so
+   * repeated reads on the same authentication share work. Synchronized so concurrent reads on
+   * different lazy fields don't double-fetch. Mapping rules are never resolved for BASIC auth — the
+   * converter does not wire the supplier and {@link #mappingRules()} returns an empty list.
    */
-  MembershipResolver newResolver(String username);
+  private final class Provider implements MembershipProvider {
+    private final EnumMap<EntityType, Set<String>> ownerTypeToIds = new EnumMap<>(EntityType.class);
+
+    private List<String> groups;
+    private List<String> roles;
+    private List<String> tenants;
+
+    Provider(final String username) {
+      ownerTypeToIds.put(EntityType.USER, Set.of(username));
+    }
+
+    @Override
+    public synchronized List<String> groups() {
+      if (groups == null) {
+        final var ids =
+            groupServices
+                .getGroupsByMemberTypeAndMemberIds(
+                    ownerTypeToIds, CamundaAuthentication.anonymous())
+                .stream()
+                .map(GroupEntity::groupId)
+                .collect(Collectors.toSet());
+        if (!ids.isEmpty()) {
+          ownerTypeToIds.put(GROUP, ids);
+        }
+        groups = List.copyOf(ids);
+      }
+      return groups;
+    }
+
+    @Override
+    public synchronized List<String> roles() {
+      if (roles == null) {
+        groups();
+
+        final var ids =
+            roleServices
+                .getRolesByMemberTypeAndMemberIds(ownerTypeToIds, CamundaAuthentication.anonymous())
+                .stream()
+                .map(RoleEntity::roleId)
+                .collect(Collectors.toSet());
+
+        if (!ids.isEmpty()) {
+          ownerTypeToIds.put(EntityType.ROLE, ids);
+        }
+        roles = List.copyOf(ids);
+      }
+      return roles;
+    }
+
+    @Override
+    public synchronized List<String> tenants() {
+      if (tenants == null) {
+        roles();
+
+        tenants =
+            tenantServices
+                .getTenantsByMemberTypeAndMemberIds(
+                    ownerTypeToIds, CamundaAuthentication.anonymous())
+                .stream()
+                .map(TenantEntity::tenantId)
+                .toList();
+      }
+      return tenants;
+    }
+
+    @Override
+    public List<String> mappingRules() {
+      return List.of();
+    }
+  }
 }
