+++ authentication/src/test/java/io/camunda/authentication/config/controllers/OidcFlowTestContext.java
@@ -8,8 +8,8 @@
 package io.camunda.authentication.config.controllers;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import io.camunda.authentication.service.MembershipService;
 import io.camunda.authentication.service.NoDbOidcMembershipService;
-import io.camunda.authentication.service.OidcMembershipService;
 import io.camunda.search.clients.auth.DisabledResourceAccessProvider;
 import io.camunda.security.api.context.CamundaAuthenticationProvider;
 import io.camunda.security.api.model.CamundaAuthentication;
@@ -94,7 +94,7 @@ public class OidcFlowTestContext {
   }
 
   @Bean
-  public OidcMembershipService createMembershipService(
+  public MembershipService createMembershipService(
       final SecurityConfiguration securityConfiguration) {
     return new NoDbOidcMembershipService(securityConfiguration);
   }
