+++ authentication/src/test/java/io/camunda/authentication/oidc/OidcBearerUserInfoClaimGapIT.java
@@ -149,7 +149,7 @@ public class OidcBearerUserInfoClaimGapIT extends AbstractWebSecurityConfigTest 
     verify(tokenClaimsConverter).convert(claimsCaptor.capture());
 
     // The customer's acceptance criterion: groups must arrive at TokenClaimsConverter so that
-    // OidcMembershipService.newResolver(...) can grant group-based authorizations. On current
+    // MembershipService.createProvider(...) can grant group-based authorizations. On current
     // code this assertion fails because only JWT claims are passed through.
     assertThat(claimsCaptor.getValue())
         .containsEntry("sub", "alice")
